Cleaning up - Client class to it's own file Client.ts - Import/Export of private key and public key to it's own file Secret.ts - Removed old unused code
